Add m1 homebrew path, hammerspoon script, fkill

+++ fish/.config/fish/functions/fkill.fish
@@ -1,6 +1,6 @@
 function fkill --description "Kill the selected processes"
 	set -l pid (ps -ef | sed 1d | fzf -m | awk '{print$2}')
-	if test 'x$pid' != 'x'
+	if test -n "$pid" && test "x$pid" != 'x'
 		kill -9 $pid
 	end
 end
